// src/interface/interface.tsx
export interface ISender {
    _id: string;
    rojifiId: string;
    creator: string;
    teams: string;
    country: string;
    countryflag: string;
    businessRegistrationNumber: string;
    businessVerificationCompleted: boolean;
    businessVerificationCompletedAt: Date | null;
    businessName: string;
    taxIdentificationNumber: string;
    taxIdentificationNumberVerified: boolean;
    taxIdentificationNumberVerifiedAt: Date | null;
    volume: number;
    email: string;
    phoneNumber: string;
    phoneCountryCode: string;
    countryOfIncorporation: string;
    percentageOwnership: number;
    affiliationStatus: AffiliationStatus;
    dateOfIncorporation: Date;
    businessAddress: string;
    businessCity: string;
    businessState: string;
    businessPostalCode: string;
    creatorFirstName: string;
    creatorLastName: string;
    creatorMiddleName: string;
    creatorDateOfBirth: Date;
    creatorPosition: string;
    creatorBirthCountry: string;
    isBeneficialOwner: BooleanString;
    creatorRole: Role;
    creatorAddress: string;
    creatorCity: string;
    creatorState: string;
    creatorPostalCode: string;
    creatorPercentageOwnership: number;
    creatorEmail: string;
    creatorIsBusinessContact: BooleanString;
    creatorVotingRightPercentage: number;
    creatorTaxIdentificationNumber: string;
    creatorTaxIdentificationNumberVerified: boolean;
    creatorTaxIdentificationNumberVerifiedAt: Date | null;
    creatorSocialSecurityNumber: string;
    creatorSocialSecurityNumberVerified: boolean;
    creatorSocialSecurityNumberVerifiedAt: Date | null;
    deleted: boolean;
    deletedAt: Date | null;
    archived: boolean;
    archivedAt: Date | null;
    businessMemorandumAndArticlesOfAssociationKyc: string;
    businessMemorandumAndArticlesOfAssociationKycVerified: boolean;
    businessMemorandumAndArticlesOfAssociationKycVerifiedAt: Date | null;
    businessCertificateOfIncorporationKyc: string;
    businessCertificateOfIncorporationKycVerified: boolean;
    businessCertificateOfIncorporationKycVerifiedAt: Date | null;
    businessCertificateOfIncorporationStatusReportKyc: string;
    businessCertificateOfIncorporationStatusReportKycVerified: boolean;
    businessCertificateOfIncorporationStatusReportKycVerifiedAt: Date | null;
    businessProofOfAddressKyc: string;
    businessProofOfAddressKycVerified: boolean;
    businessProofOfAddressKycVerifiedAt: Date | null;
    status: SenderStatus;
    createdAt: Date;
    updatedAt: Date;
}